requirements.txt
flask
openai
python-dotenv
azure-ai-openai
